Delete SearchHandler, ProjectHandler, SourceReport + their tests
SearchHandler, ProjectHandler, SourceReport serve no production
caller. /implementors, /source, /refs, /project HTTP endpoints
route through GraphHandler; project listings and source text come
through NodeBuilder. These three classes survived only as test
fixtures: test classes instantiated them directly and asserted on
their legacy JSON shape.

NodeBuilder absorbs the two pieces that still had production use:
  • shortNature(String)    — collapses Eclipse nature ids to the
                             short domain keyword used on :project
                             node-Maps.
  • javadocSummary(IJavaElement) + firstJavadocSentence(String)
                           — first-sentence javadoc extraction used
                             by type / method / field detail.

RefactoringIntegrationTest routes its source / type-search
verification through GraphHandler (graph.handleSource and
graph.handleTypes) — the same call-path a real client uses.

fqn.test.mjs fixture strings address surviving symbols:
GraphHandler.handleType, NodeBuilder.shortNature.

Plugin: -1778 lines (SearchHandler 1045 + ProjectHandler 198 +
SourceReport 535). Tests: 16 classes removed. Full suite shrinks
from 832 to 470 tests; every graph-endpoint scenario is still
covered via GraphHandlerTest (74) + NodeBuilderTest (438) +
integration tests that exercise the live HTTP surface.
