+++ plugin.tests/src/io/github/kaluchi/jdtbridge/RefactoringIntegrationTest.java
@@ -25,7 +25,7 @@ public class RefactoringIntegrationTest {
 
     private static final RefactoringHandler handler =
             new RefactoringHandler();
-    private static final SearchHandler search = new SearchHandler();
+    private static final GraphHandler graph = new GraphHandler();
 
     @BeforeAll
     public static void setUp() throws Exception {
@@ -129,8 +129,8 @@ public class RefactoringIntegrationTest {
                 ResourcesPlugin.FAMILY_AUTO_BUILD, null);
 
         // Verify caller updated
-        String source = search.handleSource(
-                Map.of("class", "test.refactor.RenameCaller"), ProjectScope.ALL).body();
+        String source = graph.handleSource(
+                Map.of("of", "test.refactor.RenameCaller"));
         assertTrue(source.contains("incrementCounter"),
                 "Caller should use new name: " + source);
     }
@@ -150,8 +150,8 @@ public class RefactoringIntegrationTest {
             // Verify getter updated
             Job.getJobManager().join(
                     ResourcesPlugin.FAMILY_AUTO_BUILD, null);
-            String source = search.handleSource(
-                    Map.of("class", "test.refactor.RenameTarget"), ProjectScope.ALL).body();
+            String source = graph.handleSource(
+                    Map.of("of", "test.refactor.RenameTarget"));
             assertTrue(source.contains("count"),
                     "Should use new field name: " + source);
         } catch (IllegalArgumentException e) {
@@ -176,13 +176,14 @@ public class RefactoringIntegrationTest {
                 ResourcesPlugin.FAMILY_AUTO_BUILD, null);
 
         // Verify new type exists
-        String findJson = search.handleFind(Map.of("name", "RenamedTarget"), ProjectScope.ALL);
+        String findJson = graph.handleTypes(
+                Map.of("pattern", "RenamedTarget"), ProjectScope.ALL);
         assertTrue(findJson.contains("test.refactor.RenamedTarget"),
                 "Should find renamed type: " + findJson);
 
         // Verify caller references updated
-        String callerSrc = search.handleSource(
-                Map.of("class", "test.refactor.RenameCaller"), ProjectScope.ALL).body();
+        String callerSrc = graph.handleSource(
+                Map.of("of", "test.refactor.RenameCaller"));
         assertTrue(callerSrc.contains("RenamedTarget"),
                 "Caller should reference RenamedTarget: " + callerSrc);
     }
@@ -203,8 +204,8 @@ public class RefactoringIntegrationTest {
                     ResourcesPlugin.FAMILY_AUTO_BUILD, null);
 
             // Verify type in new package
-            String findJson = search.handleFind(
-                    Map.of("name", "RenameCaller"), ProjectScope.ALL);
+            String findJson = graph.handleTypes(
+                    Map.of("pattern", "RenameCaller"), ProjectScope.ALL);
             assertTrue(findJson.contains("test.moved.RenameCaller"),
                     "Should be in test.moved: " + findJson);
         } catch (IllegalArgumentException e) {
